Remove mlModelCursor in favor of generic cursor table
